encode fixes

+++ contracts/modules/gamemanager/GameCreation.sol
@@ -23,7 +23,7 @@
 
 pragma solidity ^0.5.5;
 
-import '../proxy/Proxied.sol';
+import "../proxy/Proxied.sol";
 import "../datetime/TimeFrame.sol";
 import "../databases/EndowmentDB.sol";
 import "../databases/GMSetterDB.sol";
@@ -32,15 +32,15 @@ import "../../GameVarAndFee.sol";
 import "../endowment/EndowmentFund.sol";
 import "./Scheduler.sol";
 import "../../libs/SafeMath.sol";
-import '../../authority/Guard.sol';
+import "../../authority/Guard.sol";
 import "../../interfaces/IKittyCore.sol";
 import "./GameStore.sol";
 import "./GameManager.sol";
 import "../../CronJob.sol";
 import "./Forfeiter.sol";
-import '../kittieHELL/KittieHell.sol';
+import "../kittieHELL/KittieHell.sol";
 import "../endowment/HoneypotAllocationAlgo.sol";
-import '../endowment/KtyUniswap.sol';
+import "../endowment/KtyUniswap.sol";
 import "../databases/GenericDB.sol";
 import "../../withdrawPool/WithdrawPool.sol";
 import "../databases/AccountingDB.sol";
@@ -64,37 +64,62 @@ contract GameCreation is Proxied, Guard {
     AccountingDB public accountingDB;
     GameManagerHelper public gameManagerHelper;
 
-
     //EVENTS
-    event NewGame(uint indexed gameId, address playerBlack, uint kittieBlack, address playerRed, uint kittieRed, uint gameStartTime);
-    event NewListing(uint indexed kittieId, address indexed owner, uint timeListed);
+    event NewGame(
+        uint256 indexed gameId,
+        address playerBlack,
+        uint256 kittieBlack,
+        address playerRed,
+        uint256 kittieRed,
+        uint256 gameStartTime
+    );
+    event NewListing(
+        uint256 indexed kittieId,
+        address indexed owner,
+        uint256 timeListed
+    );
     // event Scheduled(uint indexed jobId, uint jobTime, uint indexed gameId, string job);
 
     //mapping(uint256 => uint256) public cronJobsForGames; //using gmGetterDB.getCronJobForGame()/gmSetterDB.setCronJobForGame() instead
 
-    modifier onlyKittyOwner(address player, uint kittieId) {
-        require(cryptoKitties.ownerOf(kittieId) == player, "Not the owner of this kittie");
+    modifier onlyKittyOwner(address player, uint256 kittieId) {
+        require(
+            cryptoKitties.ownerOf(kittieId) == player,
+            "Not the owner of this kittie"
+        );
         _;
     }
 
     /**
-    * @dev Sets related contracts
-    * @dev Can be called only by the owner of this contract
-    */
+     * @dev Sets related contracts
+     * @dev Can be called only by the owner of this contract
+     */
     function initialize() external onlyOwner {
         gmSetterDB = GMSetterDB(proxy.getContract(CONTRACT_NAME_GM_SETTER_DB));
         gmGetterDB = GMGetterDB(proxy.getContract(CONTRACT_NAME_GM_GETTER_DB));
-        endowmentFund = EndowmentFund(proxy.getContract(CONTRACT_NAME_ENDOWMENT_FUND));
-        gameVarAndFee = GameVarAndFee(proxy.getContract(CONTRACT_NAME_GAMEVARANDFEE));
+        endowmentFund = EndowmentFund(
+            proxy.getContract(CONTRACT_NAME_ENDOWMENT_FUND)
+        );
+        gameVarAndFee = GameVarAndFee(
+            proxy.getContract(CONTRACT_NAME_GAMEVARANDFEE)
+        );
         scheduler = Scheduler(proxy.getContract(CONTRACT_NAME_SCHEDULER));
-        cryptoKitties = IKittyCore(proxy.getContract(CONTRACT_NAME_CRYPTOKITTIES));
+        cryptoKitties = IKittyCore(
+            proxy.getContract(CONTRACT_NAME_CRYPTOKITTIES)
+        );
         gameStore = GameStore(proxy.getContract(CONTRACT_NAME_GAMESTORE));
         cronJob = CronJob(proxy.getContract(CONTRACT_NAME_CRONJOB));
         kittieHELL = KittieHell(proxy.getContract(CONTRACT_NAME_KITTIEHELL));
         genericDB = GenericDB(proxy.getContract(CONTRACT_NAME_GENERIC_DB));
-        withdrawPool = WithdrawPool(proxy.getContract(CONTRACT_NAME_WITHDRAW_POOL));
-        accountingDB = AccountingDB(proxy.getContract(CONTRACT_NAME_ACCOUNTING_DB));
-        gameManagerHelper = GameManagerHelper(proxy.getContract(CONTRACT_NAME_GAMEMANAGER_HELPER));
+        withdrawPool = WithdrawPool(
+            proxy.getContract(CONTRACT_NAME_WITHDRAW_POOL)
+        );
+        accountingDB = AccountingDB(
+            proxy.getContract(CONTRACT_NAME_ACCOUNTING_DB)
+        );
+        gameManagerHelper = GameManagerHelper(
+            proxy.getContract(CONTRACT_NAME_GAMEMANAGER_HELPER)
+        );
     }
 
     // /**
@@ -142,10 +167,10 @@ contract GameCreation is Proxied, Guard {
     //     onlyKittyOwner(playerBlack, kittyBlack)
     // {
     //     require(genericDB.getUintStorage(CONTRACT_NAME_TIMEFRAME, keccak256(abi.encodePacked(
-    //         genericDB.getUintStorage(CONTRACT_NAME_TIMEFRAME, keccak256(abi.encode("activeEpoch"))),"endTimeForGames"))) > gameStartTime,
+    //         genericDB.getUintStorage(CONTRACT_NAME_TIMEFRAME, keccak256(abi.encodePacked("activeEpoch"))),"endTimeForGames"))) > gameStartTime,
     //         "Wrong start time");
 
-    //     require(!genericDB.getBoolStorage(CONTRACT_NAME_SCHEDULER, keccak256(abi.encode("schedulerMode"))), "No manual match mode");
+    //     require(!genericDB.getBoolStorage(CONTRACT_NAME_SCHEDULER, keccak256(abi.encodePacked("schedulerMode"))), "No manual match mode");
 
     //     require(!scheduler.isKittyListedForMatching(kittyRed), "fighter already listed");
     //     require(!scheduler.isKittyListedForMatching(kittyBlack), "fighter already listed");
@@ -163,25 +188,34 @@ contract GameCreation is Proxied, Guard {
      * @dev Creates game and generates gameId
      * @return gameId
      */
-    function generateFight
-    (
-        address playerRed, address playerBlack,
-        uint256 kittyRed, uint256 kittyBlack,
-        uint gameStartTime
-    )
-        internal
-    {
+    function generateFight(
+        address playerRed,
+        address playerBlack,
+        uint256 kittyRed,
+        uint256 kittyBlack,
+        uint256 gameStartTime
+    ) internal {
         uint256 gameId = gmSetterDB.createGame(
-            playerRed, playerBlack, kittyRed, kittyBlack, gameStartTime);
-        
+            playerRed,
+            playerBlack,
+            kittyRed,
+            kittyBlack,
+            gameStartTime
+        );
+
         gameStore.lockVars(gameId);
 
-        (uint initialKTY, uint initialEth) = HoneypotAllocationAlgo(proxy.getContract(CONTRACT_NAME_HONEYPOT_ALLOCATION_ALGO)).generateHoneyPot(gameId);
+        (uint256 initialKTY, uint256 initialEth) = HoneypotAllocationAlgo(
+            proxy.getContract(CONTRACT_NAME_HONEYPOT_ALLOCATION_ALGO)
+        )
+            .generateHoneyPot(gameId);
         gmSetterDB.setHoneypotInfo(gameId, initialKTY, initialEth);
 
-        uint poolId = TimeFrame(proxy.getContract(CONTRACT_NAME_TIMEFRAME)).getActiveEpochID();
+        uint256 poolId = TimeFrame(proxy.getContract(CONTRACT_NAME_TIMEFRAME))
+            .getActiveEpochID();
 
-        EndowmentDB(proxy.getContract(CONTRACT_NAME_ENDOWMENT_DB)).setPoolIDinGame(gameId, poolId);
+        EndowmentDB(proxy.getContract(CONTRACT_NAME_ENDOWMENT_DB))
+            .setPoolIDinGame(gameId, poolId);
 
         gameManagerHelper.updateKittiesGame(kittyBlack, kittyRed, gameId);
 
@@ -193,23 +227,37 @@ contract GameCreation is Proxied, Guard {
 
         accountingDB.setListingFeeInGame(gameId, kittyRed, kittyBlack);
 
-        emit NewGame(gameId, playerBlack, kittyBlack, playerRed, kittyRed, gameStartTime);
+        emit NewGame(
+            gameId,
+            playerBlack,
+            kittyBlack,
+            playerRed,
+            kittyRed,
+            gameStartTime
+        );
     }
 
     /**
      * @dev External function for Scheduler to call
      * @return gameId
      */
-    function createFight
-    (
-        address playerRed, address playerBlack,
-        uint256 kittyRed, uint256 kittyBlack,
-        uint gameStartTime
+    function createFight(
+        address playerRed,
+        address playerBlack,
+        uint256 kittyRed,
+        uint256 kittyBlack,
+        uint256 gameStartTime
     )
         external
         only2Contracts(CONTRACT_NAME_SCHEDULER, CONTRACT_NAME_LIST_KITTIES)
     {
-        generateFight(playerRed, playerBlack, kittyRed, kittyBlack, gameStartTime);
+        generateFight(
+            playerRed,
+            playerBlack,
+            kittyRed,
+            kittyBlack,
+            gameStartTime
+        );
     }
 
     // function updateKitties(address winner, address loser, uint256 gameId)
@@ -222,88 +270,121 @@ contract GameCreation is Proxied, Guard {
     //     //Kill losers's Kittie
     //     kittieHELL.killKitty(gmGetterDB.getKittieInGame(gameId, loser), gameId);
 
-    //     if(genericDB.getBoolStorage(CONTRACT_NAME_SCHEDULER, keccak256(abi.encode("schedulerMode"))))
+    //     if(genericDB.getBoolStorage(CONTRACT_NAME_SCHEDULER, keccak256(abi.encodePacked("schedulerMode"))))
     //         scheduler.startGame();
     // }
 
     // ==== CRONJOBS FUNCTIONS
 
     function scheduleJobs(uint256 gameId, uint256 state)
-    external
-    onlyContract(CONTRACT_NAME_GM_SETTER_DB)
+        external
+        onlyContract(CONTRACT_NAME_GM_SETTER_DB)
     {
-        if(state == 0){
-            (,uint preStartTime,) = gmGetterDB.getGameTimes(gameId);
-            uint scheduledJob = cronJob.addCronJob(CONTRACT_NAME_GAMECREATION, preStartTime, abi.encodeWithSignature("updateGameStateCron(uint256)", gameId));
+        if (state == 0) {
+            (, uint256 preStartTime, ) = gmGetterDB.getGameTimes(gameId);
+            uint256 scheduledJob = cronJob.addCronJob(
+                CONTRACT_NAME_GAMECREATION,
+                preStartTime,
+                abi.encodeWithSignature("updateGameStateCron(uint256)", gameId)
+            );
             // emit Scheduled(scheduledJob, preStartTime, gameId, "Change state to 1");
             gmSetterDB.setCronJobForGame(gameId, scheduledJob);
         }
 
-        if(state == 1){
+        if (state == 1) {
             //If it is PRE_GAME STATE, again, when the job is scheduled (startTime), it should start, as both players press start
             //So if state did not change we must cancelGame
             //If they both press start this job is cancelled (In start function of GameManager)
-            (uint startTime,,) = gmGetterDB.getGameTimes(gameId);
-            uint scheduledJob = cronJob.addCronJob(CONTRACT_NAME_GAMECREATION, startTime, abi.encodeWithSignature("callForfeiterCron(uint256)", gameId));
+            (uint256 startTime, , ) = gmGetterDB.getGameTimes(gameId);
+            uint256 scheduledJob = cronJob.addCronJob(
+                CONTRACT_NAME_GAMECREATION,
+                startTime,
+                abi.encodeWithSignature("callForfeiterCron(uint256)", gameId)
+            );
             // emit Scheduled(scheduledJob, startTime, gameId, "Change state to 2");
             gmSetterDB.setCronJobForGame(gameId, scheduledJob);
         }
-        if(state == 2){
+        if (state == 2) {
             //If it is MAIN_GAME we endgame immediately
             //We reschedule this Job if game extends (In checkPerformance of GameManager)
-            (,,uint endTime) = gmGetterDB.getGameTimes(gameId);
-            uint scheduledJob = cronJob.addCronJob(CONTRACT_NAME_GAMECREATION, endTime, abi.encodeWithSignature("callGameEndCron(uint256)", gameId));
+            (, , uint256 endTime) = gmGetterDB.getGameTimes(gameId);
+            uint256 scheduledJob = cronJob.addCronJob(
+                CONTRACT_NAME_GAMECREATION,
+                endTime,
+                abi.encodeWithSignature("callGameEndCron(uint256)", gameId)
+            );
             // emit Scheduled(scheduledJob, endTime, gameId, "Change state to 3");
             gmSetterDB.setCronJobForGame(gameId, scheduledJob);
         }
     }
 
-    
     function updateGameStateCron(uint256 gameId)
         external
         onlyContract(CONTRACT_NAME_CRONJOB)
     {
-        uint state = gmGetterDB.getGameState(gameId);
+        uint256 state = gmGetterDB.getGameState(gameId);
         //Check forfeiter before updating game state
-        Forfeiter(proxy.getContract(CONTRACT_NAME_FORFEITER)).checkGameStatusCron(gameId, state);
+        Forfeiter(proxy.getContract(CONTRACT_NAME_FORFEITER))
+            .checkGameStatusCron(gameId, state);
         state = gmGetterDB.getGameState(gameId);
         if (state == 0) gmSetterDB.updateGameStateCron(gameId);
     }
 
-    function callForfeiterCron(uint gameId)
+    function callForfeiterCron(uint256 gameId)
         external
         onlyContract(CONTRACT_NAME_CRONJOB)
     {
-        Forfeiter(proxy.getContract(CONTRACT_NAME_FORFEITER)).forfeitCron(gameId, "Did not hit start");
+        Forfeiter(proxy.getContract(CONTRACT_NAME_FORFEITER)).forfeitCron(
+            gameId,
+            "Did not hit start"
+        );
     }
 
-    function callGameEndCron(uint gameId)
+    function callGameEndCron(uint256 gameId)
         external
         onlyContract(CONTRACT_NAME_CRONJOB)
     {
-        uint state = gmGetterDB.getGameState(gameId);
-        if (state == 2) GameManager(proxy.getContract(CONTRACT_NAME_GAMEMANAGER)).gameEndCron(gameId);
+        uint256 state = gmGetterDB.getGameState(gameId);
+        if (state == 2)
+            GameManager(proxy.getContract(CONTRACT_NAME_GAMEMANAGER))
+                .gameEndCron(gameId);
     }
 
-    function rescheduleCronJob(uint gameId)
+    function rescheduleCronJob(uint256 gameId)
         external
         onlyContract(CONTRACT_NAME_GAMEMANAGER)
     {
         //Reschedule CronJob when more time added
 
         uint256 jobId = gmGetterDB.getCronJobForGame(gameId);
-        (,,uint endTime) = gmGetterDB.getGameTimes(gameId);
-        uint newJobId = cronJob.rescheduleCronJob(CONTRACT_NAME_GAMECREATION, jobId, endTime);
-        uint activeEpoch = genericDB.getUintStorage(CONTRACT_NAME_TIMEFRAME, keccak256(abi.encodePacked("activeEpoch")));
-        if(endTime > genericDB.getUintStorage(CONTRACT_NAME_TIMEFRAME, keccak256(abi.encodePacked(activeEpoch,"restDayStart")))) {
+        (, , uint256 endTime) = gmGetterDB.getGameTimes(gameId);
+        uint256 newJobId = cronJob.rescheduleCronJob(
+            CONTRACT_NAME_GAMECREATION,
+            jobId,
+            endTime
+        );
+        uint256 activeEpoch = genericDB.getUintStorage(
+            CONTRACT_NAME_TIMEFRAME,
+            keccak256(abi.encodePacked("activeEpoch"))
+        );
+        if (
+            endTime >
+            genericDB.getUintStorage(
+                CONTRACT_NAME_TIMEFRAME,
+                keccak256(abi.encodePacked(activeEpoch, "restDayStart"))
+            )
+        ) {
             withdrawPool.addGamingDelay(endTime);
         }
         gmSetterDB.setCronJobForGame(gameId, newJobId);
     }
 
-    function deleteCronjob(uint gameId)
+    function deleteCronjob(uint256 gameId)
         external
-        only2Contracts(CONTRACT_NAME_GAMEMANAGER, CONTRACT_NAME_GAMEMANAGER_HELPER)
+        only2Contracts(
+            CONTRACT_NAME_GAMEMANAGER,
+            CONTRACT_NAME_GAMEMANAGER_HELPER
+        )
     {
         uint256 jobId = gmGetterDB.getCronJobForGame(gameId);
         cronJob.deleteCronJob(CONTRACT_NAME_GAMECREATION, jobId);
